+++ src/java/org/apache/commons/math/stat/univariate/moment/FirstMoment.java
@@ -42,7 +42,7 @@ import org.apache.commons.math.stat.univariate.AbstractStorelessUnivariateStatis
  * one of the threads invokes the <code>increment()</code> or 
  * <code>clear()</code> method, it must be synchronized externally.
  *
- * @version $Revision: 1.18 $ $Date: 2004/07/04 09:02:36 $
+ * @version $Revision: 1.19 $ $Date: 2004/07/10 17:09:08 $
  */
 public class FirstMoment extends AbstractStorelessUnivariateStatistic 
     implements Serializable {
@@ -103,7 +103,7 @@ public class FirstMoment extends AbstractStorelessUnivariateStatistic
     /**
      * @see org.apache.commons.math.stat.univariate.StorelessUnivariateStatistic#getN()
      */
-    public double getN() {
+    public long getN() {
         return n;
     }
