+++ src/java/org/apache/commons/math/stat/univariate/moment/Kurtosis.java
@@ -36,7 +36,7 @@ import org.apache.commons.math.stat.univariate.AbstractStorelessUnivariateStatis
  * one of the threads invokes the <code>increment()</code> or 
  * <code>clear()</code> method, it must be synchronized externally.
  * 
- * @version $Revision: 1.26 $ $Date: 2004/07/04 22:06:07 $
+ * @version $Revision: 1.27 $ $Date: 2004/07/10 17:09:08 $
  */
 public class Kurtosis extends AbstractStorelessUnivariateStatistic implements Serializable {
 
@@ -119,7 +119,7 @@ public class Kurtosis extends AbstractStorelessUnivariateStatistic implements Se
     /**
      * @see org.apache.commons.math.stat.univariate.StorelessUnivariateStatistic#getN()
      */
-    public double getN() {
+    public long getN() {
         return moment.getN();
     }
