+++ src/java/org/apache/commons/math/stat/univariate/moment/Mean.java
@@ -44,7 +44,7 @@ import org.apache.commons.math.stat.univariate.summary.Sum;
  * one of the threads invokes the <code>increment()</code> or 
  * <code>clear()</code> method, it must be synchronized externally.
  * 
- * @version $Revision: 1.20 $ $Date: 2004/07/04 09:02:36 $
+ * @version $Revision: 1.21 $ $Date: 2004/07/10 17:09:08 $
  */
 public class Mean extends AbstractStorelessUnivariateStatistic 
     implements Serializable {
@@ -106,7 +106,7 @@ public class Mean extends AbstractStorelessUnivariateStatistic
     /**
      * @see org.apache.commons.math.stat.univariate.StorelessUnivariateStatistic#getN()
      */
-    public double getN() {
+    public long getN() {
         return moment.getN();
     }
