+++ src/java/org/apache/commons/math/stat/univariate/moment/Variance.java
@@ -38,7 +38,7 @@ import org.apache.commons.math.stat.univariate.AbstractStorelessUnivariateStatis
  * one of the threads invokes the <code>increment()</code> or 
  * <code>clear()</code> method, it must be synchronized externally.
  * 
- * @version $Revision: 1.24 $ $Date: 2004/07/04 22:09:07 $
+ * @version $Revision: 1.25 $ $Date: 2004/07/10 17:09:08 $
  */
 public class Variance extends AbstractStorelessUnivariateStatistic implements Serializable {
 
@@ -96,7 +96,7 @@ public class Variance extends AbstractStorelessUnivariateStatistic implements Se
     /**
      * @see org.apache.commons.math.stat.univariate.StorelessUnivariateStatistic#getN()
      */
-    public double getN() {
+    public long getN() {
         return moment.getN();
     }
